Update Script
Add more data retrieval for RemoteDomains, and removing statement about it returning all domains where forwarding is allowed (as that is now a outbound spam policy setting). Gateway accounts should not need to login to everyone's mailbox, but some will have a configuration that allows the login and removal of content, though this could be done via EWS or Graph and would not be listed here. SendAs does not work as returns invalid object in CSV. Fixed.